requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
langchain-pinecone
langchain-community

